+++ R/filter.R
@@ -231,12 +231,6 @@ filter_eval <- function(
     },
     `dplyr:::signal_filter_one_column_matrix` = function(e) {
       warn_filter_one_column_matrix(env = error_call, user_env = user_env)
-    },
-    `dplyr:::signal_filter_across` = function(e) {
-      warn_filter_across(env = error_call, user_env = user_env)
-    },
-    `dplyr:::signal_filter_data_frame` = function(e) {
-      warn_filter_data_frame(env = error_call, user_env = user_env)
     }
   )
 
@@ -250,18 +244,25 @@ filter_bullets <- function(cnd, ...) {
 
 #' @export
 `filter_bullets.dplyr:::filter_incompatible_type` <- function(cnd, ...) {
-  column_name <- cnd$dplyr_error_data$column_name
   index <- cnd$dplyr_error_data$index
   result <- cnd$dplyr_error_data$result
 
-  if (is.null(column_name)) {
-    input_name <- glue("..{index}")
-  } else {
-    input_name <- glue("..{index}${column_name}")
-  }
-  glue(
-    "`{input_name}` must be a logical vector, not {obj_type_friendly(result)}."
+  bullets <- cli::format_inline(
+    "`..{index}` must be a logical vector, not {obj_type_friendly(result)}."
   )
+
+  if (is.data.frame(result)) {
+    # Provide some extra advice for people who try and use `across()` inside
+    # of `filter()`
+    bullets <- c(
+      bullets,
+      i = cli::format_inline(
+        "If you used {.fn across} to generate this data frame, please use {.fn if_any} or {.fn if_all} instead."
+      )
+    )
+  }
+
+  bullets
 }
 
 #' @export
@@ -281,29 +282,5 @@ warn_filter_one_column_matrix <- function(env, user_env) {
     env = env,
     user_env = user_env,
     always = TRUE
-  )
-}
-
-warn_filter_across <- function(env, user_env) {
-  # TODO: https://github.com/tidyverse/dplyr/issues/7758
-  lifecycle::deprecate_warn(
-    when = "1.0.8",
-    what = I("Using `across()` in `filter()`"),
-    with = I("`if_any()` or `if_all()`"),
-    always = TRUE,
-    env = env,
-    user_env = user_env
-  )
-}
-
-warn_filter_data_frame <- function(env, user_env) {
-  # TODO: https://github.com/tidyverse/dplyr/issues/7758
-  lifecycle::deprecate_warn(
-    when = "1.0.8",
-    what = I("Returning data frames from `filter()` expressions"),
-    with = I("`if_any()` or `if_all()`"),
-    always = TRUE,
-    env = env,
-    user_env = user_env
   )
 }